move check_match out of librustc
